@@ -4,16 +4,16 @@
   "author": "opensource@apollographql.com",
   "license": "MIT",
   "scripts": {
-    "postinstall": "lerna exec -- npm install --package-lock=false && lerna run prepare",
-    "build": "lerna run build",
-    "test": "jest --config ./config/jest.config.js",
-    "test:watch": "jest --config ./config/jest.config.js --watch",
-    "test:coverage": "jest --config ./config/jest.config.js --verbose --coverage",
+    "postinstall": "npx lerna exec -- npm install --package-lock=false && npx lerna run prepare",
+    "build": "npx lerna run build",
+    "test": "npx jest --config ./config/jest.config.js",
+    "test:watch": "npx jest --config ./config/jest.config.js --watch",
+    "test:coverage": "npx jest --config ./config/jest.config.js --verbose --coverage",
     "test:ci": "npm run test:coverage -- --ci --maxWorkers=2 --reporters=default --reporters=jest-junit",
-    "bundlesize": "lerna run build && bundlesize",
-    "prettier": "prettier --config ./config/prettier.config.js --write \"./**/*.{js,jsx,ts*,md,graphql,json}\"",
-    "deploy": "lerna publish --contents ./build/dist",
-    "clean": "rm -Rf ./node_modules ./meta && lerna exec -- npm run clean"
+    "bundlesize": "npx lerna run build && bundlesize",
+    "prettier": "npx prettier --config ./config/prettier.config.js --write \"./**/*.{js,jsx,ts*,md,graphql,json}\"",
+    "deploy": "npx lerna publish --contents ./build/dist",
+    "clean": "rm -Rf ./node_modules ./meta && npx lerna exec -- npm run clean"
   },
   "dependencies": {
     "@apollo/react-common": "file:./packages/common",
@@ -23,31 +23,26 @@
     "@apollo/react-testing": "file:./packages/testing"
   },
   "devDependencies": {
-    "@types/enzyme": "^3.9.1",
-    "@types/enzyme-adapter-react-16": "^1.0.5",
     "@types/graphql": "^14.2.0",
     "@types/hoist-non-react-statics": "^3.3.1",
     "@types/jest": "^24.0.11",
     "@types/react": "^16.8.13",
     "@types/react-dom": "^16.8.3",
-    "@types/react-test-renderer": "^16.8.1",
     "@types/recompose": "^0.30.5",
     "apollo-cache": "^1.2.1",
     "apollo-cache-inmemory": "^1.5.1",
     "apollo-client": "^2.5.1",
     "apollo-link": "^1.2.11",
     "bundlesize": "^0.17.1",
-    "enzyme": "^3.9.0",
-    "enzyme-adapter-react-16": "^1.12.1",
     "graphql": "^14.2.1",
     "graphql-tag": "^2.10.1",
     "jest": "^24.7.1",
     "jest-junit": "^6.3.0",
     "lerna": "^3.13.2",
     "lodash.times": "^4.3.2",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-test-renderer": "^16.8.6",
+    "react": "^16.9.0-alpha.0",
+    "react-dom": "^16.9.0-alpha.0",
+    "react-testing-library": "^7.0.0",
     "recompose": "^0.30.0",
     "rollup": "^1.9.3",
     "rollup-plugin-commonjs": "^9.3.4",
@@ -56,6 +51,7 @@
     "rollup-plugin-typescript2": "^0.20.1",
     "rollup-plugin-uglify": "^6.0.2",
     "ts-jest": "^24.0.2",
+    "tsc-watch": "^2.1.2",
     "typescript": "^3.4.3",
     "zen-observable-ts": "^0.8.18"
   },
