@@ -1,5 +1,5 @@
 import React from 'react';
-import renderer from 'react-test-renderer';
+import { render, cleanup } from 'react-testing-library';
 import gql from 'graphql-tag';
 import ApolloClient from 'apollo-client';
 import { InMemoryCache as Cache } from 'apollo-cache-inmemory';
@@ -10,7 +10,7 @@ import {
   ChildProps,
   Query,
   QueryResult,
-  DataValue,
+  DataValue
 } from '@apollo/react-components';
 import { DocumentNode } from 'graphql';
 
@@ -24,6 +24,7 @@ describe('[queries] errors', () => {
   });
   afterEach(() => {
     console.error = error;
+    cleanup();
   });
 
   // errors
@@ -40,11 +41,11 @@ describe('[queries] errors', () => {
     const data = { allPeople: { people: [{ name: 'Luke Skywalker' }] } };
     const link = mockSingleLink({
       request: { query },
-      result: { data },
+      result: { data }
     });
     const client = new ApolloClient({
       link,
-      cache: new Cache({ addTypename: false }),
+      cache: new Cache({ addTypename: false })
     });
 
     class ErrorBoundary extends React.Component {
@@ -63,12 +64,12 @@ describe('[queries] errors', () => {
       return null;
     });
 
-    renderer.create(
+    render(
       <ApolloProvider client={client}>
         <ErrorBoundary>
           <ContainerWithData />
         </ErrorBoundary>
-      </ApolloProvider>,
+      </ApolloProvider>
     );
   });
 
@@ -85,23 +86,23 @@ describe('[queries] errors', () => {
     const data = { allPeople: { people: [{ name: 'Luke Skywalker' }] } };
     const link = mockSingleLink({
       request: { query },
-      result: { data },
+      result: { data }
     });
     const client = new ApolloClient({
       link,
-      cache: new Cache({ addTypename: false }),
+      cache: new Cache({ addTypename: false })
     });
 
     const ContainerWithData = graphql(query)(() => null);
 
-    const wrapper = renderer.create(
+    const { unmount } = render(
       <ApolloProvider client={client}>
         <ContainerWithData />
-      </ApolloProvider>,
+      </ApolloProvider>
     ) as any;
 
     try {
-      wrapper.unmount();
+      unmount();
       done();
     } catch (e) {
       throw new Error(e);
@@ -120,31 +121,31 @@ describe('[queries] errors', () => {
     `;
     const link = mockSingleLink({
       request: { query },
-      error: new Error('boo'),
+      error: new Error('boo')
     });
     const client = new ApolloClient({
       link,
-      cache: new Cache({ addTypename: false }),
+      cache: new Cache({ addTypename: false })
     });
 
     const ErrorContainer = graphql(query)(
       class extends React.Component<ChildProps> {
-        componentWillReceiveProps({ data }: ChildProps) {
+        componentDidUpdate() {
+          const { data } = this.props;
           expect(data!.error).toBeTruthy();
           expect(data!.error!.networkError).toBeTruthy();
-          // expect(data.error instanceof ApolloError).toBeTruthy();
           done();
         }
         render() {
           return null;
         }
-      },
+      }
     );
 
-    renderer.create(
+    render(
       <ApolloProvider client={client}>
         <ErrorContainer />
-      </ApolloProvider>,
+      </ApolloProvider>
     );
   });
 
@@ -177,16 +178,16 @@ describe('[queries] errors', () => {
       const link = mockSingleLink(
         {
           request: { query, variables: var1 },
-          result: { data },
+          result: { data }
         },
         {
           request: { query, variables: var2 },
-          error: new Error('boo'),
-        },
+          error: new Error('boo')
+        }
       );
       const client = new ApolloClient({
         link,
-        cache: new Cache({ addTypename: false }),
+        cache: new Cache({ addTypename: false })
       });
 
       type Data = typeof data;
@@ -201,12 +202,13 @@ describe('[queries] errors', () => {
       const ErrorContainer = withState('var', 'setVar', 1)(
         graphql<Props, Data, Vars>(query)(
           class extends React.Component<ChildProps<Props, Data, Vars>> {
-            componentWillReceiveProps(props: ChildProps<Props, Data, Vars>) {
+            componentDidUpdate() {
+              const { props } = this;
               iteration += 1;
               if (iteration === 1) {
                 // initial loading state is done, we have data
                 expect(stripSymbols(props.data!.allPeople)).toEqual(
-                  data.allPeople,
+                  data.allPeople
                 );
                 props.setVar(2);
               } else if (iteration === 2) {
@@ -226,14 +228,14 @@ describe('[queries] errors', () => {
             render() {
               return null;
             }
-          },
-        ),
+          }
+        )
       );
 
-      renderer.create(
+      render(
         <ApolloProvider client={client}>
           <ErrorContainer />
-        </ApolloProvider>,
+        </ApolloProvider>
       );
     });
   });
@@ -258,11 +260,11 @@ describe('[queries] errors', () => {
 
       const link = mockSingleLink({
         request: { query },
-        error: new Error('oops'),
+        error: new Error('oops')
       });
       const client = new ApolloClient({
         link,
-        cache: new Cache({ addTypename: false }),
+        cache: new Cache({ addTypename: false })
       });
 
       const origError = console.error;
@@ -282,35 +284,27 @@ describe('[queries] errors', () => {
                 break;
               case 1:
                 // Noop. Don’t handle the error so a warning will be logged to the console.
+                expect(renderCount).toBe(2);
+                expect(errorMock.mock.calls.length).toBe(0);
+                resolve();
                 break;
               default:
                 throw new Error('Too many renders.');
             }
           } catch (error) {
-            console.error = origError;
             reject(error);
+          } finally {
+            console.error = origError;
           }
           return null;
         }
       }
 
-      renderer.create(
+      render(
         <ApolloProvider client={client}>
           <UnhandledErrorComponent />
-        </ApolloProvider>,
+        </ApolloProvider>
       );
-
-      setTimeout(() => {
-        try {
-          expect(renderCount).toBe(2);
-          expect(errorMock.mock.calls.length).toBe(0);
-          resolve();
-        } catch (error) {
-          reject(error);
-        } finally {
-          console.error = origError;
-        }
-      }, 250);
     }));
 
   it('passes any cached data when there is a GraphQL error', done => {
@@ -327,38 +321,39 @@ describe('[queries] errors', () => {
     type Data = typeof data;
     const link = mockSingleLink(
       { request: { query }, result: { data } },
-      { request: { query }, error: new Error('No Network Connection') },
+      { request: { query }, error: new Error('No Network Connection') }
     );
     const client = new ApolloClient({
       link,
-      cache: new Cache({ addTypename: false }),
+      cache: new Cache({ addTypename: false })
     });
 
     let count = 0;
     const Container = graphql<{}, Data>(query, {
-      options: { notifyOnNetworkStatusChange: true },
+      options: { notifyOnNetworkStatusChange: true }
     })(
       class extends React.Component<ChildProps<{}, Data>> {
-        componentWillReceiveProps(props: ChildProps<{}, Data>) {
+        componentDidUpdate() {
+          const { props } = this;
           try {
             switch (count++) {
               case 0:
                 expect(stripSymbols(props.data!.allPeople)).toEqual(
-                  data.allPeople,
+                  data.allPeople
                 );
                 props.data!.refetch().catch(() => null);
                 break;
               case 1:
                 expect(props.data!.loading).toBeTruthy();
                 expect(stripSymbols(props.data!.allPeople)).toEqual(
-                  data.allPeople,
+                  data.allPeople
                 );
                 break;
               case 2:
                 expect(props.data!.loading).toBeFalsy();
                 expect(props.data!.error).toBeTruthy();
                 expect(stripSymbols(props.data!.allPeople)).toEqual(
-                  data.allPeople,
+                  data.allPeople
                 );
                 done();
                 break;
@@ -373,13 +368,13 @@ describe('[queries] errors', () => {
         render() {
           return null;
         }
-      },
+      }
     );
 
-    renderer.create(
+    render(
       <ApolloProvider client={client}>
         <Container />
-      </ApolloProvider>,
+      </ApolloProvider>
     );
   });
 
@@ -400,20 +395,21 @@ describe('[queries] errors', () => {
     const link = mockSingleLink(
       { request: { query }, result: { data } },
       { request: { query }, error: new Error('This is an error!') },
-      { request: { query }, result: { data: dataTwo } },
+      { request: { query }, result: { data: dataTwo } }
     );
     const client = new ApolloClient({
       link,
-      cache: new Cache({ addTypename: false }),
+      cache: new Cache({ addTypename: false })
     });
 
     let count = 0;
     const noop = () => null;
     const Container = graphql<{}, Data>(query, {
-      options: { notifyOnNetworkStatusChange: true },
+      options: { notifyOnNetworkStatusChange: true }
     })(
       class extends React.Component<ChildProps<{}, Data>> {
-        componentWillReceiveProps(props: ChildProps<{}, Data>) {
+        componentDidUpdate() {
+          const { props } = this;
           try {
             switch (count++) {
               case 0:
@@ -437,16 +433,11 @@ describe('[queries] errors', () => {
                     done.fail('Expected good data on second refetch.');
                   });
                 break;
-              // Further fix required in QueryManager
-              // case 3:
-              //   expect(props.data.loading).toBeTruthy();
-              //   expect(props.data.error).toBeFalsy();
-              //   break;
               case 3:
                 expect(props.data!.loading).toBeFalsy();
                 expect(props.data!.error).toBeFalsy();
                 expect(stripSymbols(props.data!.allPeople)).toEqual(
-                  dataTwo.allPeople,
+                  dataTwo.allPeople
                 );
                 done();
                 break;
@@ -461,15 +452,16 @@ describe('[queries] errors', () => {
         render() {
           return null;
         }
-      },
+      }
     );
 
-    renderer.create(
+    render(
       <ApolloProvider client={client}>
         <Container />
-      </ApolloProvider>,
+      </ApolloProvider>
     );
   });
+
   it('does not throw/console.err an error after a component that received a network error is unmounted', done => {
     const query: DocumentNode = gql`
       query somethingelse {
@@ -485,12 +477,12 @@ describe('[queries] errors', () => {
     type Data = typeof data;
     const link = mockSingleLink(
       { request: { query }, result: { data } },
-      { request: { query }, error: new Error('This is an error!') },
+      { request: { query }, error: new Error('This is an error!') }
     );
 
     const client = new ApolloClient({
       link,
-      cache: new Cache({ addTypename: false }),
+      cache: new Cache({ addTypename: false })
     });
     let count = 0;
     const noop = () => null;
@@ -511,13 +503,14 @@ describe('[queries] errors', () => {
         props: something => {
           return {
             data: something.data!,
-            hideContainer: something!.ownProps.hideContainer,
+            hideContainer: something!.ownProps.hideContainer
           };
-        },
-      },
+        }
+      }
     )(
       class extends React.Component<ChildProps<QueryChildProps, Data>> {
-        componentWillReceiveProps(props: ChildProps<QueryChildProps, Data>) {
+        componentDidUpdate() {
+          const { props } = this;
           try {
             switch (count++) {
               case 0:
@@ -553,19 +546,19 @@ describe('[queries] errors', () => {
         render() {
           return null;
         }
-      },
+      }
     );
 
     class Switcher extends React.Component<any, any> {
       constructor(props: any) {
         super(props);
         this.state = {
-          showContainer: true,
+          showContainer: true
         };
       }
       render() {
         const {
-          state: { showContainer },
+          state: { showContainer }
         } = this;
         if (showContainer) {
           return (
@@ -578,12 +571,13 @@ describe('[queries] errors', () => {
       }
     }
 
-    renderer.create(
+    render(
       <ApolloProvider client={client}>
         <Switcher />
-      </ApolloProvider>,
+      </ApolloProvider>
     );
   });
+
   it('correctly sets loading state on remount after a network error', done => {
     const query: DocumentNode = gql`
       query somethingelse {
@@ -600,52 +594,85 @@ describe('[queries] errors', () => {
     type Data = typeof data;
     const link = mockSingleLink(
       { request: { query }, error: new Error('This is an error!') },
-      { request: { query }, result: { data: dataTwo } },
+      { request: { query }, result: { data: dataTwo } }
     );
     const client = new ApolloClient({
       link,
-      cache: new Cache({ addTypename: false }),
+      cache: new Cache({ addTypename: false })
     });
 
     let count = 0;
     type ContainerOwnProps = { toggle: () => void };
     const Container = graphql<ContainerOwnProps, Data>(query, {
-      options: { notifyOnNetworkStatusChange: true },
+      options: { notifyOnNetworkStatusChange: true }
     })(
       class extends React.Component<ChildProps<ContainerOwnProps, Data>> {
-        componentWillReceiveProps(props: ChildProps<ContainerOwnProps, Data>) {
+        componentDidUpdate() {
           // first payload with error has arrived from server
-          switch (count++) {
+          // switch (count) {
+          //   case 0:
+          //     count = 1;
+          //     expect(this.props.data!.error!.networkError!.message).toMatch(
+          //       /This is an error/
+          //     );
+          //     done();
+          //     // unmount this component
+          //     // this.props.toggle();
+          //     // setTimeout(() => {
+          //     //   // remount after 50 ms
+          //     //   this.props.toggle();
+          //     // }, 50);
+          //     break;
+          //   case 1:
+          //     // data has arrived
+          //     expect(this.props.data!.loading).toBe(false);
+          //     done();
+          //     break;
+          //   default:
+          //     throw new Error('Too many renders.');
+          // }
+        }
+
+        componentWillUnmount() {
+          // unmount after first error
+          // expect(count).toBe(1);
+        }
+
+        render() {
+          switch (count) {
             case 0:
-              expect(props.data!.error!.networkError!.message).toMatch(
-                /This is an error/,
+              expect(this.props.data!.loading).toBe(true);
+              break;
+            case 1:
+              expect(this.props.data!.loading).toBe(false);
+              expect(this.props.data!.error!.networkError!.message).toMatch(
+                /This is an error/
               );
               // unmount this component
-              this.props.toggle();
+              setTimeout(() => {
+                this.props.toggle();
+              }, 0);
               setTimeout(() => {
                 // remount after 50 ms
                 this.props.toggle();
               }, 50);
               break;
-            case 1:
+            case 2:
+              expect(this.props.data!.loading).toBe(true);
+              break;
+            case 3:
               // data has arrived
-              expect(props.data!.loading).toBe(false);
+              expect(this.props.data!.loading).toBe(false);
               done();
               break;
             default:
               throw new Error('Too many renders.');
           }
-        }
+          count += 1;
 
-        componentWillUnmount() {
-          // unmount after first error
-          expect(count).toBe(1);
-        }
-
-        render() {
           return null;
         }
-      },
+      }
     );
 
     type Toggle = () => void;
@@ -658,17 +685,18 @@ describe('[queries] errors', () => {
       render() {
         if (!this.state.show) return null;
         return this.props.children(() =>
-          this.setState(({ show }) => ({ show: !show })),
+          this.setState(({ show }) => ({ show: !show }))
         );
       }
     }
 
-    renderer.create(
+    render(
       <ApolloProvider client={client}>
         <Manager>{(toggle: Toggle) => <Container toggle={toggle} />}</Manager>
-      </ApolloProvider>,
+      </ApolloProvider>
     );
   });
+
   describe('errorPolicy', () => {
     it('passes any GraphQL errors in props along with data', done => {
       const query: DocumentNode = gql`
@@ -685,26 +713,27 @@ describe('[queries] errors', () => {
         result: {
           data: {
             allPeople: {
-              people: null,
-            },
+              people: null
+            }
           },
-          errors: [new Error('this is an error')],
-        },
+          errors: [new Error('this is an error')]
+        }
       });
 
       const client = new ApolloClient({
         link,
-        cache: new Cache({ addTypename: false }),
+        cache: new Cache({ addTypename: false })
       });
 
       const ErrorContainer = graphql(query, {
-        options: { errorPolicy: 'all' },
+        options: { errorPolicy: 'all' }
       })(
         class extends React.Component<ChildProps> {
-          componentWillReceiveProps({ data }: ChildProps) {
+          componentDidUpdate() {
+            const { data } = this.props;
             expect(data!.error).toBeTruthy();
             expect(data!.error!.graphQLErrors[0].message).toEqual(
-              'this is an error',
+              'this is an error'
             );
             expect(data).toMatchObject({ allPeople: { people: null } });
             done();
@@ -712,15 +741,16 @@ describe('[queries] errors', () => {
           render() {
             return null;
           }
-        },
+        }
       );
 
-      renderer.create(
+      render(
         <ApolloProvider client={client}>
           <ErrorContainer />
-        </ApolloProvider>,
+        </ApolloProvider>
       );
     });
+
     it('passes any GraphQL errors in props along with data [component]', done => {
       const query: DocumentNode = gql`
         query people {
@@ -736,23 +766,24 @@ describe('[queries] errors', () => {
         result: {
           data: {
             allPeople: {
-              people: null,
-            },
+              people: null
+            }
           },
-          errors: [new Error('this is an error')],
-        },
+          errors: [new Error('this is an error')]
+        }
       });
 
       const client = new ApolloClient({
         link,
-        cache: new Cache({ addTypename: false }),
+        cache: new Cache({ addTypename: false })
       });
 
       class ErrorContainer extends React.Component<QueryResult> {
-        componentWillReceiveProps(props: QueryResult) {
+        componentDidUpdate() {
+          const { props } = this;
           expect(props.error).toBeTruthy();
           expect(props.error!.graphQLErrors[0].message).toEqual(
-            'this is an error',
+            'this is an error'
           );
           expect(props.data!.allPeople!).toMatchObject({ people: null });
           done();
@@ -762,12 +793,12 @@ describe('[queries] errors', () => {
         }
       }
 
-      renderer.create(
+      render(
         <ApolloProvider client={client}>
           <Query query={query} errorPolicy="all">
             {(props: any) => <ErrorContainer {...props} />}
           </Query>
-        </ApolloProvider>,
+        </ApolloProvider>
       );
     });
   });
