@@ -1,5 +1,5 @@
 import React from 'react';
-import renderer from 'react-test-renderer';
+import { render, cleanup } from 'react-testing-library';
 import gql from 'graphql-tag';
 import ApolloClient from 'apollo-client';
 import { InMemoryCache as Cache } from 'apollo-cache-inmemory';
@@ -11,14 +11,18 @@ import { graphql } from '../../graphql';
 
 describe('[queries] polling', () => {
   let error: typeof console.error;
+
   beforeEach(() => {
     error = console.error;
     console.error = jest.fn(() => {});
     jest.useRealTimers();
   });
+
   afterEach(() => {
     console.error = error;
+    cleanup();
   });
+
   // polling
   it('allows a polling query to be created', done => {
     expect.assertions(4);
@@ -39,19 +43,19 @@ describe('[queries] polling', () => {
     const link = mockSingleLink(
       { request: { query }, result: { data } },
       { request: { query }, result: { data: data2 } },
-      { request: { query }, result: { data } },
+      { request: { query }, result: { data } }
     );
     const client = new ApolloClient({
       link,
-      cache: new Cache({ addTypename: false }),
+      cache: new Cache({ addTypename: false })
     });
 
     let count = 0;
     const Container = graphql(query, {
       options: () => ({
         pollInterval: POLL_INTERVAL,
-        notifyOnNetworkStatusChange: false,
-      }),
+        notifyOnNetworkStatusChange: false
+      })
     })(() => {
       count++;
       expect(true).toBe(true);
@@ -61,10 +65,10 @@ describe('[queries] polling', () => {
       return null;
     });
 
-    const wrapper = renderer.create(
+    render(
       <ApolloProvider client={client}>
         <Container />
-      </ApolloProvider>,
+      </ApolloProvider>
     );
   });
 
@@ -80,16 +84,17 @@ describe('[queries] polling', () => {
     `;
     const link = mockSingleLink({
       request: { query },
-      result: { data: { allPeople: { people: [{ name: 'Luke Skywalker' }] } } },
+      result: { data: { allPeople: { people: [{ name: 'Luke Skywalker' }] } } }
     });
     const client = new ApolloClient({
       link,
-      cache: new Cache({ addTypename: false }),
+      cache: new Cache({ addTypename: false })
     });
 
     const Container = graphql(query)(
       class extends React.Component<ChildProps> {
-        componentWillReceiveProps({ data }: ChildProps) {
+        componentDidUpdate() {
+          const { data } = this.props;
           expect(data!.stopPolling).toBeTruthy();
           expect(data!.stopPolling instanceof Function).toBeTruthy();
           expect(data!.stopPolling).not.toThrow();
@@ -98,12 +103,12 @@ describe('[queries] polling', () => {
         render() {
           return null;
         }
-      },
+      }
     );
-    renderer.create(
+    render(
       <ApolloProvider client={client}>
         <Container />
-      </ApolloProvider>,
+      </ApolloProvider>
     );
   });
 
@@ -119,35 +124,34 @@ describe('[queries] polling', () => {
     `;
     const link = mockSingleLink({
       request: { query },
-      result: { data: { allPeople: { people: [{ name: 'Luke Skywalker' }] } } },
+      result: { data: { allPeople: { people: [{ name: 'Luke Skywalker' }] } } }
     });
     const client = new ApolloClient({
       link,
-      cache: new Cache({ addTypename: false }),
+      cache: new Cache({ addTypename: false })
     });
-    let wrapper: renderer.ReactTestRenderer;
     const Container = graphql(query, { options: { pollInterval: 10 } })(
       class extends React.Component<ChildProps> {
-        componentWillReceiveProps({ data }: ChildProps) {
+        componentDidUpdate() {
+          const { data } = this.props;
           expect(data!.startPolling).toBeTruthy();
           expect(data!.startPolling instanceof Function).toBeTruthy();
           // XXX this does throw because of no pollInterval
           // expect(data.startPolling).not.toThrow();
           setTimeout(() => {
-            wrapper.unmount();
             done();
           }, 0);
         }
         render() {
           return null;
         }
-      },
+      }
     );
 
-    wrapper = renderer.create(
+    render(
       <ApolloProvider client={client}>
         <Container />
-      </ApolloProvider>,
+      </ApolloProvider>
     );
   });
 });
