@@ -1,6 +1,5 @@
-import React from 'react';
-import TestUtils from 'react-dom/test-utils';
-import { shallow, mount } from 'enzyme';
+import React, { useContext } from 'react';
+import { render, cleanup } from 'react-testing-library';
 import ApolloClient from 'apollo-client';
 import { InMemoryCache as Cache } from 'apollo-cache-inmemory';
 import { ApolloLink } from 'apollo-link';
@@ -9,9 +8,11 @@ import { ApolloProvider } from '../ApolloProvider';
 import { getApolloContext } from '../ApolloContext';
 
 describe('<ApolloProvider /> Component', () => {
+  afterEach(cleanup);
+
   const client = new ApolloClient({
     cache: new Cache(),
-    link: new ApolloLink((o, f) => (f ? f(o) : null)),
+    link: new ApolloLink((o, f) => (f ? f(o) : null))
   });
 
   class Child extends React.Component<any, { store: any; client: any }> {
@@ -38,7 +39,7 @@ describe('<ApolloProvider /> Component', () => {
 
     componentDidMount() {
       this.setState({
-        client: this.props.client,
+        client: this.props.client
       });
     }
 
@@ -52,23 +53,23 @@ describe('<ApolloProvider /> Component', () => {
   }
 
   it('should render children components', () => {
-    const wrapper = mount(
+    const { getByText } = render(
       <ApolloProvider client={client}>
-        <div className="unique" />
-      </ApolloProvider>,
+        <div className="unique">Test</div>
+      </ApolloProvider>
     );
 
-    expect(wrapper.contains(<div className="unique" />)).toBeTruthy();
+    expect(getByText('Test')).toBeTruthy();
   });
 
   it('should support the 2.0', () => {
-    const wrapper = mount(
+    const { getByText } = render(
       <ApolloProvider client={{} as ApolloClient<any>}>
-        <div className="unique" />
-      </ApolloProvider>,
+        <div className="unique">Test</div>
+      </ApolloProvider>
     );
 
-    expect(wrapper.contains(<div className="unique" />)).toBeTruthy();
+    expect(getByText('Test')).toBeTruthy();
   });
 
   it('should require a client', () => {
@@ -80,72 +81,66 @@ describe('<ApolloProvider /> Component', () => {
       // Before testing `ApolloProvider`, we first fully reset the
       // existing context using `ApolloContext.Provider` directly.
       const ApolloContext = getApolloContext();
-      mount(
+      render(
         <ApolloContext.Provider value={{}}>
           <ApolloProvider client={undefined as any}>
             <div className="unique" />
           </ApolloProvider>
-        </ApolloContext.Provider>,
+        </ApolloContext.Provider>
       );
     }).toThrowError(
       'ApolloProvider was not passed a client instance. Make ' +
-        'sure you pass in your client via the "client" prop.',
+        'sure you pass in your client via the "client" prop.'
     );
     console.error = originalConsoleError;
   });
 
   it('should not require a store', () => {
-    const wrapper = mount(
+    const { getByText } = render(
       <ApolloProvider client={client}>
-        <div className="unique" />
-      </ApolloProvider>,
+        <div className="unique">Test</div>
+      </ApolloProvider>
     );
-    expect(wrapper.contains(<div className="unique" />)).toBeTruthy();
+    expect(getByText('Test')).toBeTruthy();
   });
 
   it('should add the client to the children context', () => {
-    const wrapper = mount(
+    const TestChild = () => {
+      const context = useContext(getApolloContext());
+      expect(context.client).toEqual(client);
+      return null;
+    };
+    render(
       <ApolloProvider client={client}>
-        <Child />
-        <Child />
-      </ApolloProvider>,
+        <TestChild />
+        <TestChild />
+      </ApolloProvider>
     );
-    const children = wrapper.find(Child);
-    expect(children).toHaveLength(2);
-    children.forEach(node => {
-      expect(node.instance().context.client).toEqual(client);
-    });
   });
 
   it('should update props when the client changes', () => {
-    const container = shallow(<Container client={client} />);
-    expect(container.find(ApolloProvider).props().client).toEqual(client);
+    let clientToCheck = client;
+
+    const TestChild = () => {
+      const context = useContext(getApolloContext());
+      expect(context.client).toEqual(clientToCheck);
+      return null;
+    };
+    const { rerender } = render(
+      <ApolloProvider client={clientToCheck}>
+        <TestChild />
+      </ApolloProvider>
+    );
 
     const newClient = new ApolloClient({
       cache: new Cache(),
-      link: new ApolloLink((o, f) => (f ? f(o) : null)),
+      link: new ApolloLink((o, f) => (f ? f(o) : null))
     });
-    container.setState({ client: newClient });
-    expect(container.find(ApolloProvider).props().client).toEqual(newClient);
-    expect(container.find(ApolloProvider).props().client).not.toEqual(client);
-  });
-
-  it('child component should be able to query new client when props change', () => {
-    const container = TestUtils.renderIntoDocument<React.Component<any, any>>(
-      <Container client={client} />,
-    ) as React.Component<any, any>;
-
-    const child = TestUtils.findRenderedComponentWithType(container, Child);
-    expect(child.context.client).toEqual(client);
-
-    const newClient = new ApolloClient({
-      cache: new Cache(),
-      link: new ApolloLink((o, f) => (f ? f(o) : null)),
-    });
-
-    container.setState({ client: newClient });
-
-    expect(child.context.client).toEqual(newClient);
-    expect(child.context.client).not.toEqual(client);
+    clientToCheck = newClient;
+    rerender(
+      <ApolloProvider client={clientToCheck}>
+        <TestChild />
+      </ApolloProvider>
+    );
   });
 });
